Restructured directory layout of included libraries
