+++ stag/spectra-1.0.1/Spectra/GenEigsBase.h
