+++ stag/spectra-1.0.1/Spectra/MatOp/SparseGenMatProd.h
