+++ stag/spectra-1.0.1/Spectra/MatOp/SymShiftInvert.h
