+++ stag/spectra-1.0.1/Spectra/SymEigsBase.h
