+++ stag/spectra-1.0.1/Spectra/Util/CompInfo.h
