+++ stag/spectra-1.0.1/Spectra/Util/SelectionRule.h
