+++ stag/spectra-1.0.1/Spectra/Util/SimpleRandom.h
