+++ stag/spectra-1.0.1/Spectra/Util/Version.h
